+++ addon/-private/execution_context/acceptance-native-dom.js
@@ -1,10 +1,18 @@
-import ExecutionContext from './context';
+import ExecutionContext from './native-dom-context';
+
+import {
+  visit
+} from 'ember-native-dom-helpers';
 
 export default function AcceptanceExecutionContext(pageObjectNode) {
   ExecutionContext.call(this, pageObjectNode);
 }
 
 AcceptanceExecutionContext.prototype = Object.create(ExecutionContext.prototype);
+
+AcceptanceExecutionContext.prototype.visit = function() {
+  return visit(...arguments);
+};
 
 AcceptanceExecutionContext.prototype.runAsync = function(cb) {
   window.wait().then(() => {
